feat(android): Add secure live updates to Android chore(android): Reorganize code
