+++ android/build.gradle
@@ -126,7 +126,7 @@ dependencies {
     //noinspection GradleDynamicVersion
     api "io.ionic:portals:0.6.+"
     //noinspection GradleDynamicVersion
-    api "io.ionic:liveupdates:0.0.+"
+    api "io.ionic:liveupdates:0.1.+"
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.0"
 }
