+++ android/src/main/java/io/ionic/portals/reactnative/ReactNativePortalsPackage.kt
@@ -11,7 +11,6 @@ class ReactNativePortalsPackage : ReactPackage {
         return listOf(
             PortalManagerModule(reactContext),
             PortalsPubSubModule(reactContext),
-            LiveUpdatesModule(reactContext)
         )
     }
